requirements.txt
Pillow
piexif
geopy
tqdm
customtkinter